Fix editor height for Firefox
From Firefox 94 the editor no longer fills the page. I resolved this by setting the height of all elements to 100% all the way from the html node to the iframe node. To make sure the iframe does not get a scrollbar we need to set its display style to block.